[REFACTOR] refactoring the data model
