+++ src/data/models/__init__.py
@@ -1,0 +1,15 @@
+from .protrend_database import Effector
+from .protrend_database import Evidence
+from .protrend_database import Gene
+from .protrend_database import Operon
+from .protrend_database import Organism
+from .protrend_database import Pathway
+from .protrend_database import Publication
+from .protrend_database import Regulator
+from .protrend_database import RegulatoryFamily
+from .protrend_database import RegulatoryInteraction
+from .protrend_database import Source
+from .protrend_database import TFBS
+
+from .protrend_community import RegulatorCommunity
+from .protrend_community import GeneCommunity
