+++ src/data/models/protrend_database/__init__.py
@@ -1,0 +1,12 @@
+from .effector import Effector
+from .evidence import Evidence
+from .gene import Gene
+from .operon import Operon
+from .organism import Organism
+from .pathway import Pathway
+from .publication import Publication
+from .regulator import Regulator
+from .regulatory_family import RegulatoryFamily
+from .regulatory_interaction import RegulatoryInteraction
+from .source import Source
+from .tfbs import TFBS
