+++ src/data/models/protrend_database/effector.py
