+++ src/data/models/protrend_database/gene.py
