+++ src/data/models/protrend_database/operon.py
