+++ src/data/models/protrend_database/publication.py
