+++ src/data/models/protrend_database/regulatory_interaction.py
