+++ src/domain/dpi/__init__.py
@@ -7,20 +7,4 @@ from .queries import (
     get_object,
     get_objects,
     get_query_set)
-from .operations import (
-    create_objects,
-    create_relationship,
-    create_unique_relationship,
-    create_unique_reverse_relationship,
-    delete_objects,
-    delete_relationship,
-    delete_relationships,
-    get_rel_query_set,
-    get_relationship,
-    get_relationships,
-    is_connected,
-    object_validation,
-    update_objects,
-    update_relationship
-)
 from .utils import (protrend_id_decoder, protrend_id_encoder)
