+++ src/interfaces/api/serializers/organism.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from constants import help_text
-from data import Organism
+from data.models import Organism
 from interfaces.serializers.base import BaseSerializer
 from interfaces.serializers.fields import SourceField, URLField
 from interfaces.serializers.relationship import RelationshipSerializer, SourceRelationshipSerializer
