+++ src/interfaces/api/serializers/regulatory_family.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from constants import help_text, choices
-from data import RegulatoryFamily
+from data.models import RegulatoryFamily
 from interfaces.serializers.base import BaseSerializer
 from interfaces.serializers.fields import URLField, SourceField
 from interfaces.serializers.relationship import SourceRelationshipSerializer, RelationshipSerializer
