+++ src/interfaces/api/serializers/regulatory_interaction.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from constants import help_text, choices
-from data import RegulatoryInteraction
+from data.models import RegulatoryInteraction
 from interfaces.serializers.base import BaseSerializer
 from interfaces.serializers.fields import URLField, SourceField
 from interfaces.serializers.relationship import SourceRelationshipSerializer, RelationshipSerializer
